% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterimEvaluation.R
\name{InterimEvaluation}
\alias{InterimEvaluation}
\title{Interim Evaluation Package}
\usage{
InterimEvaluation(dataTrain, dataTest, saveCSV = FALSE)
}
\arguments{
\item{dataTrain}{Training dataset: a data frame or a matrix where rows are observations and columns are features with the last column being response.}

\item{dataTest}{Test dataset: a data frame or a matrix with new observations and the same set of features as in the training set apart from the output.}

\item{saveCSV}{Set to FALSE by default; otherwise saves the predictions as csv file.}
}
\value{
A list with submit dataframe containing predictions for the test data.
}
\description{
Implement random forest classifier with 400 trees for Predicting online news popularity Kaggle competition.
}

